Integrated `download` function into `server.js` Remove new function from `Homebrew.Model.js` Remove `sourceFunctions.jsx` module
