+++ server/homebrew.model.js
@@ -34,16 +34,6 @@ HomebrewSchema.methods.sanatize = function(full=false){
 	return brew;
 };
 
-HomebrewSchema.methods.escapeTextForHtmlDisplay = function(){
-	const replaceStrings = { '&': '&amp;', '<': '&lt;', '>': '&gt;' };
-	text = this.text;
-	for (const replaceStr in replaceStrings) {
-		text = text.replaceAll(replaceStr, replaceStrings[replaceStr]);
-	}
-	text = `<code><pre style="white-space: pre-wrap;">${text}</pre></code>`;
-	return text;
-};
-
 HomebrewSchema.methods.increaseView = async function(){
 	this.lastViewed = new Date();
 	this.views = this.views + 1;
